reorganise app in files and folders strucrture, #16
