+++ lib/models/task.dart
@@ -1,0 +1,8 @@
+/// Task model
+/// A task contains a string text and a status completed.
+class Task {
+  final String text;
+  bool completed;
+
+  Task({this.text, this.completed = false});
+}
