+++ test/widget_test.dart
@@ -6,8 +6,7 @@
 // tree, read text, and verify that the values of widget properties are correct.
 
 import 'package:flutter_test/flutter_test.dart';
-
-import 'package:todolist/main.dart';
+import '../lib/models/task.dart';
 
 void main() {
   test('Item created has completed to false', () {
